initial update of the bootstrapping process for post-merge.
